requirements.txt
elasticsearch
requests
tqdm
python-dotenv
fastapi
uvicorn
